Rename kind package to docker
