+++ docker/actions/cluster_actions.go
@@ -25,8 +25,8 @@ import (
 	"github.com/pkg/errors"
 	k8sversion "k8s.io/apimachinery/pkg/util/version"
 	"sigs.k8s.io/cluster-api-provider-docker/cloudinit"
-	constkind "sigs.k8s.io/cluster-api-provider-docker/kind/constants"
-	"sigs.k8s.io/cluster-api-provider-docker/kind/kubeadm"
+	constkind "sigs.k8s.io/cluster-api-provider-docker/docker/constants"
+	"sigs.k8s.io/cluster-api-provider-docker/docker/kubeadm"
 	"sigs.k8s.io/cluster-api-provider-docker/third_party/forked/loadbalancer"
 	"sigs.k8s.io/kind/pkg/cluster/constants"
 	"sigs.k8s.io/kind/pkg/cluster/nodes"
