+++ docker/actions/kind.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
-	constkind "sigs.k8s.io/cluster-api-provider-docker/kind/constants"
+	constkind "sigs.k8s.io/cluster-api-provider-docker/docker/constants"
 	"sigs.k8s.io/cluster-api-provider-docker/third_party/forked/loadbalancer"
 	"sigs.k8s.io/kind/pkg/cluster/config/defaults"
 	"sigs.k8s.io/kind/pkg/cluster/constants"
